requirements.txt
# WoldVirtual3DlucIA - Dependencias Principales
# Inteligencia Artificial Modular Estándar v0.6.0

# Core Dependencies
asyncio-mqtt==0.16.1
aiohttp==3.9.1
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.1

# Machine Learning & AI
torch==2.1.0
transformers==4.35.2
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# Database & Storage
sqlite3
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.0

# Web & API
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
requests==2.31.0
httpx==0.25.2

# Security & Authentication
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
passlib==1.7.4

# Monitoring & Logging
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0

# Data Processing
pydantic==2.5.0
marshmallow==3.20.1
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Optional Dependencies
# Para funcionalidades avanzadas
tensorflow==2.15.0
keras==2.15.0
plotly==5.17.0
dash==2.14.2
streamlit==1.28.1

# Para procesamiento de audio
librosa==0.10.1
soundfile==0.12.1

# Para procesamiento de imágenes
opencv-python==4.8.1.78
scikit-image==0.22.0

# Para networking P2P
libp2p==0.1.0
p2pnetwork==1.0.0

# Para blockchain (opcional)
web3==6.11.3
eth-account==0.9.0

# Para WebXR (opcional)
three.js==0.158.0
aframe==1.4.0
